completed README

+++ code/src/clean.sh
@@ -1,5 +1,13 @@
 #!/usr/bin/env sh
-
-rm featureSet.txt
-rm train.txt
-rm test.txt
+if [ -f featureSet.txt ];
+then
+  rm featureSet.txt
+fi
+if [ -f train.txt ];
+then
+  rm train.txt
+fi
+if [ -f test.txt ];
+then
+  rm test.txt
+fi
